requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
google-adk==1.20.0
litellm==1.81.16
requests==2.32.5
feedparser==6.0.12
